Moved tests location to __tests__
